// frontend/src/adapter/api/models/destination.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-fwe-ss-24
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface Destination
 */
export interface Destination {

    /**
     * @type {any}
     * @memberof Destination
     */
    id: any;

    /**
     * @type {any}
     * @memberof Destination
     */
    name: any;

    /**
     * @type {any}
     * @memberof Destination
     */
    description?: any;

    /**
     * @type {any}
     * @memberof Destination
     */
    startDate: any;

    /**
     * @type {any}
     * @memberof Destination
     */
    endDate: any;

    /**
     * @type {any}
     * @memberof Destination
     */
    activities: any;

    /**
     * @type {any}
     * @memberof Destination
     */
    photos?: any;

    /**
     * @type {any}
     * @memberof Destination
     */
    createdAt: any;

    /**
     * @type {any}
     * @memberof Destination
     */
    updatedAt: any;

    /**
     * @type {any}
     * @memberof Destination
     */
    rating?: any;

    /**
     * @type {any}
     * @memberof Destination
     */
    trips: any;
}